# api/openapi3_v1.yaml
...
openapi: 3.0.3
info:
  title: Uniplay private API
  version: 1.0.0
servers:
- url: http://localhost:8080/v1
...
paths:
  /players/{steam_id}:
    get:
      tags:
      - player
      summary: Профиль игрока
      operationId: getPlayerProfile
      parameters:
      - name: steam_id
        in: path
        description: steam id игрока
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: uint64
          x-go-type: uint64
          x-go-name: SteamID
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /replays:
    post:
      tags:
      - replay
      summary: Загрузить запись матча
      operationId: uploadReplay
      requestBody:
        $ref: '#/components/requestBodies/ReplayUploadRequestBody'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      required:
      - code
      - message
      type: object
      properties:
        code:
          type: integer
          description: статус ответа или код ошибки >= 600
        message:
          type: string
          description: сообщение ошибки
    Player:
      required:
      - create_time
      - steam_id
      - team_flag_code
      - team_name
      - update_time
      type: object
      properties:
        steam_id:
          type: integer
          format: uint64
          x-go-name: SteamID
          x-go-type: uint64
        team_name:
          type: string
        team_flag_code:
          type: string
        create_time:
          type: string
          description: RFC3339 datetime string
          format: date-time
          x-go-type: time.Time
        update_time:
          type: string
          description: RFC3339 datetime string
          format: date-time
          x-go-type: time.Time
...
    MatchTeam:
      required:
      - clan_name
      - flag_code
      - player_steam_ids
      - score
      type: object
      properties:
        clan_name:
          type: string
          description: название команды, берется из загруженной записи матча
        flag_code:
          type: string
          description: ISO 3166 код флага
        score:
          type: integer
          format: int8
          x-go-type: int8
        player_steam_ids:
          type: array
          items:
            type: integer
            format: uint64
            x-go-type: uint64
    Match:
      required:
      - map_name
      - match_duration
      - match_id
      - team1
      - team2
      - uploaded_at
      type: object
      properties:
        match_id:
          type: string
          format: uuid
          x-go-name: MatchID
          x-go-type: google_uuid.UUID
          x-go-type-import:
            name: google_uuid
            path: github.com/google/uuid
        map_name:
          type: string
        match_duration:
          type: integer
          format: int64
          x-go-type: time.Duration
        team1:
          $ref: '#/components/schemas/MatchTeam'
        team2:
          $ref: '#/components/schemas/MatchTeam'
        uploaded_at:
          type: string
          description: datetime RFC3339
          format: date-time
          x-go-type: time.Time
...
  requestBodies:
    ReplayUploadRequestBody:
      content:
        multipart/form-data:
          schema:
            type: object
            properties:
              replay:
                type: string
                description: запись матча в формате .dem, макс. размер 300 мб
                format: binary
      required: true
...